Petstagram start
I just initialized our project in this folder. All of the client side files are created in /public folder and more client side files will be created in there in the future as well. All of the server side files should be outside of the /public folder. Good Luck guys! This is the start of our 3 weeks journey.